requirements.txt
# Core
flask==3.0.0
werkzeug==3.0.1
python-dotenv==1.0.0

# Image Processing
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scikit-image==0.22.0
scipy==1.11.4

# PDF Processing
pdf2image==1.16.3
img2pdf==0.5.0

# Utilities
python-magic==0.4.27

# Production
gunicorn==21.2.0
gevent==23.9.1

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
